casting instances where printf format %llu is used to unsigned long long to silence warning on gcc

+++ Core/SoarKernel/src/print.cpp
@@ -231,7 +231,8 @@ char *symbol_to_string (agent* thisAgent, Symbol *sym,
 	  dest=thisAgent->printed_output_string;
 	  dest_size = MAX_LEXEME_LENGTH*2+10; /* from agent.h */
 	}
-    SNPRINTF (dest, dest_size, "%c%llu", sym->id.name_letter, sym->id.name_number);
+    // BADBAD: static casting for llu portability
+    SNPRINTF (dest, dest_size, "%c%llu", sym->id.name_letter, static_cast<unsigned long long>(sym->id.name_number));
 	dest[dest_size - 1] = 0; /* ensure null termination */
     return dest;
